- Separate HomePage into multiple component (Dropzone and Filelist)
